work on macros
